created chart table

+++ ProfitLibrary/GraphData.cs
@@ -5,7 +5,7 @@ namespace ProfitLibrary
     public class GraphData
     {
         public List<string> XTicks { get; set; }
-        public long YTicks { get; set; }
+        public double YTicks { get; set; }
         public List<Point> Points { get; set; }
     }
 }
